# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  telegramId: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  superadmin: "superadmin",
  head_office_admin: "head_office_admin",
  editor: "editor",
  branch_head: "branch_head",
  hunter: "hunter",
} as const;

export interface Branch {
  id: number;
  name: string;
  city: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface User {
  id: number;
  telegramId: string;
  name: string;
  role: UserRole;
  branchId?: number | null;
  branch?: Branch | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export interface CreateBranchBody {
  name: string;
  city: string;
  isActive?: boolean;
}

export interface UpdateBranchBody {
  name?: string;
  city?: string;
  isActive?: boolean;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  superadmin: "superadmin",
  head_office_admin: "head_office_admin",
  editor: "editor",
  branch_head: "branch_head",
  hunter: "hunter",
} as const;

export interface CreateUserBody {
  telegramId: string;
  name: string;
  role: CreateUserBodyRole;
  branchId?: number | null;
  password: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  superadmin: "superadmin",
  head_office_admin: "head_office_admin",
  editor: "editor",
  branch_head: "branch_head",
  hunter: "hunter",
} as const;

export interface UpdateUserBody {
  name?: string;
  role?: UpdateUserBodyRole;
  branchId?: number | null;
  isActive?: boolean;
  password?: string;
}

export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus];

export const ClientStatus = {
  draft: "draft",
  questionnaire: "questionnaire",
  recommendation: "recommendation",
  basket: "basket",
  pdf_generated: "pdf_generated",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface Client {
  id: number;
  sessionId: string;
  fullName?: string | null;
  phone?: string | null;
  status: ClientStatus;
  branchId: number;
  branch?: Branch | null;
  assignedToId?: number | null;
  assignedTo?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface ClientListResponse {
  data: Client[];
  total: number;
  page: number;
  pageSize: number;
}

export interface CreateClientBody {
  fullName?: string;
  phone?: string;
  branchId: number;
  assignedToId?: number;
}

export type UpdateClientBodyStatus =
  (typeof UpdateClientBodyStatus)[keyof typeof UpdateClientBodyStatus];

export const UpdateClientBodyStatus = {
  draft: "draft",
  questionnaire: "questionnaire",
  recommendation: "recommendation",
  basket: "basket",
  pdf_generated: "pdf_generated",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface UpdateClientBody {
  fullName?: string;
  phone?: string;
  status?: UpdateClientBodyStatus;
  assignedToId?: number;
}

export interface ProductCategory {
  id: number;
  name: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateProductCategoryBody {
  name: string;
  description?: string;
}

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

export const ProductType = {
  credit: "credit",
  non_credit: "non_credit",
} as const;

export interface Product {
  id: number;
  name: string;
  type: ProductType;
  categoryId?: number | null;
  category?: ProductCategory | null;
  description?: string | null;
  minAmount?: number | null;
  maxAmount?: number | null;
  minTermMonths?: number | null;
  maxTermMonths?: number | null;
  interestRate?: number | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateProductBodyType =
  (typeof CreateProductBodyType)[keyof typeof CreateProductBodyType];

export const CreateProductBodyType = {
  credit: "credit",
  non_credit: "non_credit",
} as const;

export interface CreateProductBody {
  name: string;
  type: CreateProductBodyType;
  categoryId?: number;
  description?: string;
  minAmount?: number;
  maxAmount?: number;
  minTermMonths?: number;
  maxTermMonths?: number;
  interestRate?: number;
  isActive?: boolean;
}

export type UpdateProductBodyType =
  (typeof UpdateProductBodyType)[keyof typeof UpdateProductBodyType];

export const UpdateProductBodyType = {
  credit: "credit",
  non_credit: "non_credit",
} as const;

export interface UpdateProductBody {
  name?: string;
  type?: UpdateProductBodyType;
  categoryId?: number;
  description?: string;
  minAmount?: number;
  maxAmount?: number;
  minTermMonths?: number;
  maxTermMonths?: number;
  interestRate?: number;
  isActive?: boolean;
}

export interface Article {
  id: number;
  title: string;
  content: string;
  isPublished: boolean;
  targetAllBranches: boolean;
  branchIds: number[];
  authorId?: number | null;
  author?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateArticleBody {
  title: string;
  content: string;
  isPublished?: boolean;
  targetAllBranches?: boolean;
  branchIds?: number[];
}

export interface UpdateArticleBody {
  title?: string;
  content?: string;
  isPublished?: boolean;
  targetAllBranches?: boolean;
  branchIds?: number[];
}

export interface DashboardSummary {
  totalClients: number;
  totalActiveClients: number;
  totalCompletedToday: number;
  totalBranches: number;
  totalHunters: number;
  totalProducts: number;
  completedThisMonth: number;
  rejectedThisMonth: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  entityId?: number | null;
  entityType?: string | null;
  userId?: number | null;
  userName?: string | null;
  branchName?: string | null;
  createdAt: string;
}

export interface BranchStat {
  branchId: number;
  branchName: string;
  totalClients: number;
  completedClients: number;
  activeHunters: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ListUsersParams = {
  branchId?: number;
  role?: string;
  isActive?: boolean;
};

export type ListClientsParams = {
  branchId?: number;
  assignedTo?: number;
  status?: string;
  search?: string;
  page?: number;
  pageSize?: number;
};

export type ListProductsParams = {
  categoryId?: number;
  isActive?: boolean;
  type?: string;
};

export type ListArticlesParams = {
  branchId?: number;
  isPublished?: boolean;
  search?: string;
};

export type GetRecentActivityParams = {
  limit?: number;
};