# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minerva Credit Hunter API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with Telegram ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBranchBody"
      responses:
        "200":
          description: Branch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
        - name: isPublished
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Article updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        telegramId:
          type: string
        password:
          type: string
      required: [telegramId, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, city, isActive, createdAt, updatedAt]
    CreateBranchBody:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        isActive:
          type: boolean
      required: [name, city]
    UpdateBranchBody:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        isActive:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [superadmin, head_office_admin, editor, branch_head, hunter]
        branchId:
          type: integer
          nullable: true
        branch:
          $ref: "#/components/schemas/Branch"
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, telegramId, name, role, isActive, createdAt, updatedAt]
...
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
        targetAllBranches:
          type: boolean
        branchIds:
          type: array
          items:
            type: integer
        authorId:
          type: integer
          nullable: true
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, content, isPublished, targetAllBranches, branchIds, createdAt, updatedAt]
    CreateArticleBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
        targetAllBranches:
          type: boolean
        branchIds:
          type: array
          items:
            type: integer
      required: [title, content]
    UpdateArticleBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
        targetAllBranches:
          type: boolean
        branchIds:
          type: array
          items:
            type: integer
...