# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  name: string;
  city: string;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}