# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArticleBody {
  title: string;
  content: string;
  isPublished?: boolean;
  targetAllBranches?: boolean;
  branchIds?: number[];
}