# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minerva Credit Hunter API
 * OpenAPI spec version: 0.1.0
 */
import type { Branch } from "./branch";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  telegramId: string;
  name: string;
  role: UserRole;
  branchId?: number | null;
  branch?: Branch | null;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}